+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [zim] `get_libkiwix_humanid()` to get kiwix-serve BookName from a filename
+
+### Changed
+
+- [packages] `ZimPackage.get_url()` now uses libkiwix's human ID
+
 ## [1.11.0] - 2024-02-17
 
 ### Added